Adding multivariant support for Identifiable extensions
Support for Connectable extensions through AbstractMultiVariantConnectableExtension class is deprecated, as Connectable extends Identifiable
