+++ iidm/iidm-impl/src/main/java/com/powsybl/iidm/network/impl/AbstractMultiVariantConnectableExtension.java
@@ -6,31 +6,21 @@
  */
 package com.powsybl.iidm.network.impl;
 
-import com.powsybl.commons.exceptions.UncheckedClassCastExceptionException;
-import com.powsybl.commons.extensions.AbstractExtension;
 import com.powsybl.iidm.network.Connectable;
-import com.powsybl.iidm.network.Network;
 
 /**
+ * @deprecated use {@link AbstractMultiVariantIdentifiableExtension} instead.
  * @author Ghiles Abdellah <ghiles.abdellah at rte-france.com>
  */
-public abstract class AbstractMultiVariantConnectableExtension<T extends Connectable> extends AbstractExtension<T> implements MultiVariantObject {
+@Deprecated
+public abstract class AbstractMultiVariantConnectableExtension<T extends Connectable<T>> extends AbstractMultiVariantIdentifiableExtension<T> {
 
+    /**
+     * Deprecated, use {@link AbstractMultiVariantIdentifiableExtension} instead.
+     */
+    @Deprecated
     public AbstractMultiVariantConnectableExtension(T extendable) {
         super(extendable);
     }
 
-    protected VariantManagerHolder getVariantManagerHolder() {
-        Network network = getExtendable().getNetwork();
-
-        if (network instanceof VariantManagerHolder) {
-            return (VariantManagerHolder) network;
-        }
-
-        throw new UncheckedClassCastExceptionException("network cannot be converted to VariantManagerHolder");
-    }
-
-    protected int getVariantIndex() {
-        return getVariantManagerHolder().getVariantIndex();
-    }
 }
